Exercise management and Exercise and User Management

+++ implementation/webapp/model/UserModel.php
@@ -1,6 +1,7 @@
 <?php
 
     require_once($_SERVER['DOCUMENT_ROOT']."/honours/webapp/model/Model.php");
+    require_once($_SERVER['DOCUMENT_ROOT']."/honours/webapp/model/PermissionLevels.php");
 
     class UserModel extends Model
     {
@@ -63,18 +64,102 @@
             return parent::retrieve(json_encode($variables, JSON_INVALID_UTF8_SUBSTITUTE), $paramTypes);
         }
 
+        //get a users mark for an exercise
+        public function getExerciseMark($userId, $codeId)
+        {
+            //get the total marks available
+            $this->sqlStmt = 'SELECT count(codeId_fk) as total FROM honours_code_answer WHERE codeId_fk = ?';
+
+            $variables = new \stdClass();
+            $variables -> codeId_fk = $codeId;
+
+            $paramTypes = "i";
+
+            $totalJson = parent::retrieve(json_encode($variables, JSON_INVALID_UTF8_SUBSTITUTE), $paramTypes);
+
+            if (!$totalJson)
+            {
+                return null;
+            }
+
+            $total = json_decode($totalJson, JSON_INVALID_UTF8_SUBSTITUTE);
+            $totalPoints = 0;
+
+            if (!isset($total["isempty"]))
+            {
+                $totalPoints = $total[0]["total"];
+            }
+
+            //get user points for this exercise
+            $this->sqlStmt = 'SELECT mark FROM honours_user_exercise WHERE userId = ? AND codeId = ?';
+
+            $variables = new \stdClass();
+            $variables -> userId = $userId;
+            $variables -> codeId = $codeId;
+
+            $paramTypes = "ii";
+
+            $pointsJson = parent::retrieve(json_encode($variables, JSON_INVALID_UTF8_SUBSTITUTE), $paramTypes);
+
+            if (!$pointsJson)
+            {
+                return null;
+            }
+
+            $points = json_decode($pointsJson, JSON_INVALID_UTF8_SUBSTITUTE);
+            $userPoints = 0;
+
+            if (!isset($points["isempty"]))
+            {
+                $userPoints = $points[0]["mark"];
+            }
+            
+            $ret = array("points" => $userPoints, "total" => $totalPoints);
+
+            return json_encode($ret);
+
+        }
+
         //delete a user
         public function deleteData($jsonData)
         {
             //get the primary key passed through json data
             $data = json_decode($jsonData, JSON_INVALID_UTF8_SUBSTITUTE|JSON_OBJECT_AS_ARRAY);
 
-            $this->sqlStmt = 'DELETE FROM honours_user WHERE userId = ?';
+            //get user data
+            $this->sqlStmt = 'SELECT permissionLevel FROM honours_user WHERE userId = ?';
 
             $WHERE_variables = new \stdClass();
             $WHERE_variables->userId = $data["userId"];
 
             $paramTypes = "i";
+
+            $userJson = parent::retrieve(json_encode($WHERE_variables, JSON_INVALID_UTF8_SUBSTITUTE), $paramTypes);
+
+            //if we fail to get user data, return false
+            if (!$userJson)
+            {
+                return false;
+            }
+
+            $user = json_decode($userJson, JSON_INVALID_UTF8_SUBSTITUTE);
+
+            //if the user doesnt exist, return true
+            if (isset($user["isempty"]))
+            {
+                return true;
+            }
+
+            $permission = $user[0]["permissionLevel"];
+
+            //check if this is an admin user, if so, do not delete
+            if ($permission == PermissionLevels::ADMIN)
+            {
+                return false;
+            }
+
+            //delete user
+            $this->sqlStmt = 'DELETE FROM honours_user WHERE userId = ?';
 
             return parent::delete(json_encode($WHERE_variables), $paramTypes);
         }
@@ -101,7 +186,7 @@
 
             $paramTypes = "siii";
 
-            return parent::create($jsonData, $paramTypes);
+            return parent::update($jsonData, $paramTypes);
         }
 
         //login user
